modify controller

+++ src/controller/include/controller/comp_error_core.h
@@ -8,18 +8,18 @@
 #define COMP_ERROR
 
 #include <ros/ros.h>
-#include <geometry_msgs/Pose2D.h>
 #include <geometry_msgs/TwistStamped.h>
 #include <path_follower/Trajectory2D.h>
 #include <path_follower/state_Dynamic.h>
+#include <controller/TrackingInfo.h>
 
 #include <path_follower/Vehicle.h>
 #include <math.h>
 #include <vector>
 using namespace std;
 
-double getDistance(const path_follower::TrajectoryPoint2D point, const path_follower::state_Dynamic state, double ra_to_cg, double ds);
+float getDistance(const path_follower::TrajectoryPoint2D point, const path_follower::state_Dynamic state, float ra_to_cg, float ds);
 
-vector<double> ComputeTrackingError(const path_follower::Trajectory2D traj, const path_follower::state_Dynamic state, double ra_to_cg, double ds);
+vector<float> ComputeTrackingError(const path_follower::Trajectory2D traj, const path_follower::state_Dynamic state, float ra_to_cg, float ds);
 
 #endif
